Return EGL_FALSE if eglChooseConfig failed to find any matching configs
Technically, we shouldn't have touched *num_config at all in this
case, but the remote interface to the host doesn't return a separate
boolean return value, only the number of matching configs.

This makes sure that callers that only check the return value of
eglChooseConfig notice a failure to find configs.

Change-Id: I7f660a4a5b8306184608a86c34329d3163ad05ba