Fixes interpretation of attribute lists in the EGL emulation library.
In a few places, EGL_NONE was incorrectly assumed to be 0. EGL_NONE is
used to indicate the end of the list. By assuming EGL_NONE is 0, the
code would fail to detect the end of the list appropriately.

Bug: 23036813

(cherry picked from commit 6801005a9e9e5741cbc73a98d7282485767d7e96)

Change-Id: I909d48a15a2cebfda93fb165e55821d7b694fa4b