More gracefulness on dequeueBuffer_* fail on flip
BUG: https://code.google.com/p/android/issues/detail?id=204176

Features like Instant Run will do things like abandon
the current set of buffers SurfaceFlinger gave to the UI,
which makes functions like eglSwapBuffers fail,
since they are dequeueing a buffer that is considered 'invalid'
(as it is before the instant run).

If we set the EGL error to EGL_BAD_ALLOC, this interacts badly
with libs/hwui/renderthread/EGLManager.cpp, causing
a LOG_ALWAYS_FATAL(...) to be called, which aborts the current process.

This CL sets the EGL error to EGL_BAD_SURFACE instead,
which allows hwui to recover.

Change-Id: I7cb5bf3b3834b3696981746f2319804c1adf6a25