Temporarily rename makefile to avoid building duplicate opengl libs.
We've yet to drop the opengl code from the old location. Turns out that
simply including this new source location leads the Android build system
to build the lib from both the old and new location.

Renamed the Android.mk file here to avoid building from this location
until the migration is complete.

Change-Id: I66b5f4201f99b8a17c814918e5b42489d686d7c5