Create pre-process gralloc pipe on the guest
https://buganizer.corp.google.com/issues/29457657

This patch create a pre-process pipe in gralloc. It tells the host which
process holds which color buffer, so that on process exit the host can
correctly reduce the reference counters of all gralloc color buffers.

Each process will get a 64bit identifier from the host, to help the host
identifying it.

It works with host patch:
https://android-review.googlesource.com/#/c/246823/

Change-Id: I8d9a512a1f151f4f29a3b318ccf17d62d8f46512