Goldfish sync driver user-space interface
goldfish_sync.h implements the ioctls that interface
with the goldfish sync driver and the goldfish sync
virtual device on the host.

goldfish_sync_open() tells the kernel driver that
there is a new context in which sync work
may be required. This is represented by a user-space
FD. Note that this is not a timeline FD
as with SW_SYNC_USER.

The intention is to associate each EGL context
with a sync context, reflected in getGoldfishSyncFd()
in the EGLContext_t class.

The IOC_QUEUE_WORK ioctl may return a valid fence FD
that can be used by the system to do work while
waiting for that fence FD to be signaled.

In order to avoid bad behavior, timelines and sync fence objects
are still completely maintained and signaled by the
kernel driver and underlying host system.

This is part of a sequential, multi-CL change. There is also
a corresponding multi-CL change on the host side:

https://android-review.googlesource.com/#/q/topic:emu-glsync-host

In particular, the ioctl() implementation is listed in

https://android-review.googlesource.com/238399

The changes in the system image are as follows:

platform/build:

https://googleplex-android-review.git.corp.google.com/1024926

device/generic/goldfish:

https://googleplex-android-review.git.corp.google.com/1230942

device/generic/goldfish-opengl:

https://googleplex-android-review.git.corp.google.com/1219535
https://googleplex-android-review.git.corp.google.com/1219536
https://googleplex-android-review.git.corp.google.com/1219537 <- this CL
https://googleplex-android-review.git.corp.google.com/1219538
https://googleplex-android-review.git.corp.google.com/1219539
https://googleplex-android-review.git.corp.google.com/1219570 <- uses this
https://googleplex-android-review.git.corp.google.com/1219571

Change-Id: I74b92415acd95b4bf2ec8aa3a6ff5392d86cf0b3