Add backing implementation of EGLSyncKHR objects
For EGL_KHR_fence_sync and related extensions in the emulator,
we'll need an object that can track host-side sync objects
along with any Android sync framework fence fd's that
may be produced, and be able to maintain the type and status
of sync objects (generally, EGL fence vs. EGL fence + fence fd,
and signaled vs. unsignaled).

This is part of a sequential, multi-CL change. There is also
a corresponding multi-CL change on the host side:

https://android-review.googlesource.com/#/q/topic:emu-glsync-host

The changes in the system image are as follows:

platform/build:

https://googleplex-android-review.git.corp.google.com/1024926

device/generic/goldfish:

https://googleplex-android-review.git.corp.google.com/1230942

device/generic/goldfish-opengl:

https://googleplex-android-review.git.corp.google.com/1219535
https://googleplex-android-review.git.corp.google.com/1219536
https://googleplex-android-review.git.corp.google.com/1219537
https://googleplex-android-review.git.corp.google.com/1219538
https://googleplex-android-review.git.corp.google.com/1219539 <- this CL
https://googleplex-android-review.git.corp.google.com/1219570 <- uses this
https://googleplex-android-review.git.corp.google.com/1219571

Change-Id: Ic20d11500f3a6e7164c1a5919937b26a885b1987