Emulator: Support YV12 format in gralloc
The YV12 format is used by codec when it generates output.
Need to support it to pass android.media.cts.ImageReaderDecoderTest

Note:
This YV12 format and other yuv formats are only meant to be used
as software read/write buffer; they are not supported by the
host GL's colorbuffer. so when we want to render it to the host
screen, we have to convert to RGB565 format or other RGB format.

Bug: 30112142
Change-Id: I874e8f2e040fd15ad02ce48c0d65bed1cd08a791