Allow YUV_420_888 through gralloc
This is currently only used to help pass tests that
involve the direct use of the YUV_420_888 format.

An attempt is made at format conversion using the same
code for converting YV12. This would keep memory structures
intact, but result in the wrong colors.

The test:

android.hardware.camera2.cts.AllocationTest#testAllocationFromCameraFlexibleYuv

has a much higher probability of passing with this change,
provided the emulated camera is used,
and the other CLs:

Guest:

https://googleplex-android-review.git.corp.google.com/#/c/1282093/
https://googleplex-android-review.git.corp.google.com/#/c/1291418/

Host:

https://android-review.googlesource.com/#/c/253893/

Change-Id: I70f86402ace6debae731a7f5b98833be3096fdd6