Per-process EGLImage resource release
https://buganizer.corp.google.com/issues/29457657

This CL marks each EGLImage with its process unique ID when
communicating with the host, so that the host can properly clean up
EGLImages when the guest process is killed.

It is doing similar thing as in
https://googleplex-android-review.git.corp.google.com/#/c/1207896/

The implementation reuses gralloc pipe. Because now the gralloc pipe
is also used outside of gralloc, this patch moves the gralloc pipe
implemenation ProcessPipe.cpp and also renames gralloc pipe.

Change-Id: I90c6d1d842ba2741d371f02a519e0fa904bc0f08