[GLESv3][VAO] enable/pointer for vertex attribs when possible
If we don't forward enables/pointers to the host until
right before the draw, when a nonzero VAO is bound,
it defeats the purpose of VAO usage for performance.

This CL prepares the GLES2 encoder for VAO usage by
forwarding those calls whenever possible.

It gives a flag to sendVertexAttributes to not do any
changes in GL state if all attributes are backed by VBOs.

The reason we don't forward them 100% is because of
supporting either legacy or tricky behavior:

Legacy: specifying vertex arrays from host memory directly
with a host ptr passed to glVertexAttribPointer
Tricky: Using glVertexAttrib*f*(...) instead of
glVertexAttribPointer

Change-Id: I190aab86d7dcd7e01cfac0a1f60b44fac131abd0