Align goldfish_dma_context to 8 byte boundaries
bug: 35465006

It looks like when Chrome is started, we are in some strange world
where sizeof(goldfish_dma_context) = 16 instead of 24,
and we get in trouble with gralloc_register buffer,
and Chrome fails to boot.

TODO: Note that youtube videos inside chrome still don't seem to work,
for some reason (this doesn't depend on whether or not goldfish_dma is
used).

Change-Id: Ibc66885434b885993025b70d957ea8e1adf524de