Revise gralloc memory mgmt for O
bug: 37101765
bug: 37202533
bug: 37273615

This is a first cut at better gralloc memory management.

First, to fix CTS media tests NXing, we
track # refs for a gralloc buffer per process, and only
unmap the ashmem region if the refcount reached zero.

To allow usage with HIDL gralloc, we need to extend this
refcounting scheme to host color buffers.

Then, it's possible for goldfish_dma to pin too much memory.
(too much here meaning, being > 256 MB).

Previously, a separate phys contig region was allocated
for every gralloc buffer. This is unnnecessary, as the
goldfish_dma regions are only used to transfer data to/from the
host for lock/unlock (TODO: do it for lock).
This setup no longer holds in Treble since the guest
likes to hold on to handles for long times, which means
that the gralloc buffers themselves are much longer-lived.
ashmem can take the hit since it presumably pages to disk
and doesn't use phys contig memory, but goldfish_dma cannot
as it is limited to ~256 MB by kernel CMA.

This CL starts with some memory management for that:
it is based on only one user of the goldfish_dma at a time per
process, so only enough memory for the biggest buffer
to process is neded to be allocated in goldfish_dma.

Indeed, we then have a dedicated DMA region per process, whose
ref count is all gralloc buffers alloced or registered
in the current process; when the ref count reaches zero,
we essentially reset the max value.

This makes goldfish_dma mem usage fairly close to being constantly
24MB during the course of a android.media.cts.VideoEncoderTest,
which does lots of gralloc buffer allocations, registering,
hanging on to native_handle_t's for long times, etc.

Change-Id: Ie1aea80989f1b26c99c2603264f59a1d77b525f9