Fix O-CTS dEQP-EGL get_frame_timestamps failures
bug: 37100138

1. Support the new enum in eglSurfaceAttrib/eglQuerySurface
2. Real fix was previous CL to revert setting of async mode on surfaces:

https://googleplex-android-review.git.corp.google.com/#/c/2154275/

but that did not fix it immediately because new versions
of the test started asking for a new enum and failing
because we didn't know about the enum.

Then after the test is truly activated by us knowing about the new enum,
the test runs for real and passes (by running frameworks/native's EGL frame
timestamps getter).

Change-Id: Icff716acd0cf68689d7aaad6ccec978dc1415497