Proper fix for HAL_PIXEL_FORMAT_RGBX_8888
This is an appropriate fix for HAL_PIXEL_FORMAT_RGBX_8888.

The handling of RGBX_8888 is very subtle. Most of time we want it to
be treated as RGBA_8888, with the exception that alpha is always
ignored and treated as 1. The solution is to create 3 channel RGB
texture on the host while telling the guest to use 4 channel RGBA for
read/write. The host GL will handle the Alpha channel approriately.

BUG: 78602661

This cl does not impact real devices.

Test: atest CtsNativeHardwareTestCases
Change-Id: Ife299bfc55ce327d0213616b9902ba537cfb0d82