Fix RAM ballooning in dEQP-EGL.functional.query_context.*
bug: 111008879

When we delete the current context, we need to wait
until it is not current anymore.

The previous handling of this "delete the pending-delete context when
it is not current" never actually deleted anything, because
~EGLContext_t::IS_CURRENT was not applied yet.

This CL moves the attempt to delete pending after ~EGLContext_t:IS_CURRENT
is applied.

Test: Run dEQP-EGL.functional.query_context.* ; resident memory
should not increase by 4 GB

Change-Id: I9ab597aca698e05c166db3881591cd0843c8bd23