Autogenerate the CMake build
Adds a GNUMakefile that will generate the cmake files based upon the
android makefiles. This will make it easier to keep the makefiles in
sync. If the makefiles change, we will have to run make in this
directory.

It relies on the following:

- The android.mk files are using the emugl dependency mechanism
- You are in the emulator repo (as we need the emulator build.mk files)
- We translate the emugl -> json representation.
- We write the json representation to disk
- We have a python script that transforms the json to cmake

It is all driven by GNUmake, so all you have to do is type make to
regenerate the files. SOONG does not use GNUMakefiles so we should be
able to co-exist.

Change-Id: I6f9ae22397bcf9987699b946aa14921b00fc94b0