[vulkan] add cereal to goldfish
bug: 111137294

This merely builds guest side marshaling library and frontend library,
without hooking them up to anything yet.

cereal is a generator of Vulkan sources to make it easier to deal with
the repetitive iteration over struct fields and API arguments.

To re-generate them, check out the emu-master-dev repo and re-run
./android/scripts/generate-vulkan-sources.sh

Change-Id: I1fb42128c1099dd03aa245c7c329f5970d82adcd