[vulkan] Revise to using CMake for host and copy to guest
bug: 111137294

This CL brings the generated makefiles in line with host side CMake
while also generating guest-side Android.mk (which when building for
host, is generated back to CMake by the transformer :)

When generating guest side marshaling code, the AHB related extensions
do not need to be marshaled and will cause compile errors when used on
host, so just skip those altogether for now.

+ Refactor cerealgenerator.py to not mix up so many
types of statements together.

Change-Id: I14a3c878a435717e98b9f9e6797c2239216ba8c8