[vulkan] Integrate with HostConnection
bug: 111137294

This pushes vulkan components downward so that we can reuse
HostConnection and the general infrastructure around it (incl. DMA) when
running Vulkan.

This has benefits, like being able to interoperate easily with
renderControl and gralloc, and not having to create qemu pipe (or some
other transport) ourselves.

- Move VulkanStream to vulkan_enc
- Move cereal as a dependency of vulkan_enc

Note that this makes VulkanStream into a class that doesnt do any real
writing or reading, but that's fine since this is an intermediate state
and we'll hook it up to IOStream later on.

Change-Id: I7dce6798f1490f7a5874da3a4213d7b5bdf71f06