[vulkan] Generate classes to track resources...
...through subclasses of VulkanHandleMapping, and consolidate them into
a single global ResourceTracker class, which is also automatically
generated.

bug: 111137294
bug: 119104304
bug: 119157982

The guest needs to be able to track and wrap Vulkan handles (objects) as
they are created and destroyed.

Most of the time, this is a trivial transformation; wrapping up the
Vulkan handle into an object that is tracked in the guest, and adding
HWVULKAN_DISPATCH_MAGIC to make the Android Vulkan loader happy when
encountering Vulkan dispatchable handles.

However, sometimes, we need to create a custom Vulkan class by
ourselves. This is especially needed for VkDeviceMemory as we need to
interface with our custom implementation of allocation/map/unmap/etc.

In the future, we also have need to custom wrap other Vulkan objects,
especially if we want to support allocation callbacks properly.

So this CL sets up classes where, for every Vulkan handle type of
concern, there is a go-to method to create/destroy them, and also to
unwrap the host-side handle when streaming them to the host.

This CL also adds the deepcopy and handlemap generators to the guest,
since they will be needed to use the wrappers.

Change-Id: Ibed52eb3dbe0f7809b3c030cd94327a4d7d67ce9