[vulkan] ABI independences: fuse marshaling with handle mapping
bug: 119685647
bug: 111137294

The current way of storing underlying Vulkan host handles as the same
Vulkan handle types is not going to work if we are cross-ABI. We not
only need to store host Vulkan handles as uint64_t, but also make sure
the unmarshaled Vulkan structs that hold Vulkan handles also contain the
correct values. There is no real good way to do this other than to fuse
handle mapping into marshaling, where if a Vulkan handle inside a Vulkan
struct is encountered, we write that out as a uint64_t, then on the
host, read the uint64_t and convert it to a host Vulkan handle type.

Also:

- Refactor VulkanHandleMapping to use macros to define classes, remove
resourcetracker.py code generator

This CL does not introduce ABI independence, retaining existing
functionality while introducing handle mapping functionality to
marshaling.

Change-Id: Ifa34dabb94b648c320b5c53187dfac004a6b3ab2