[vulkan] ABI independence of Vulkan handles
bug: 111137294
bug: 119685647

This CL switches the encoder over to handle mismatches on guest
versus host of sizeof Vulkan handles.

This requires us to use the unwrap handle mapper fused with the
marshaler whenever writing anything to the host, and requires
the fusion of the create handle mapper when reading output handles
from the host.

Change-Id: I609ce45409fb4f0dea3b28fa8b1b08b75bba96c2