[vulkan] Generate code for custom logic as event handling
bug: 111137294

Called VkEventHandler, it provides functions that can be inserted
as mini-layers in the encoder or decoder implementation.

This makes it easier to add logic to the encoder/decoder.

VkEncoder will then use a VkEventHandler before encoding as a
validation, while using a VkEventHandler after encoding postprocess to
update tracking of Vulkan objects.

Change-Id: I303a72878387dd22fe8f7f1a797de67c8b3f951b