[vulkan] Add annotations for device memory virtualization
bug: 121420031

We need to virtualize VkDeviceMemory objects in the guest; however,
there are many Vulkan API calls and structs that require transformation
back to the "real" VkDeviceMemory object.

This CL adds annotations to the XML where these can happen:

devicememoryhandle
devicememoryoffset
devicememorysize
devicememorytypeindex
devicememorytypebits

Each such annotation key corresponds to a value that is the ID of the
syntactically abstract VkDeviceMemory field; this is to be able to
associate the offset, typeindex, size, etc. fields with the same
VkDeviceMemory field. While being slightly harder to work with than
"len" attributes, but duplicates less information.

Also, in the future, if a struct or API call contains references to two
different VkDeviceMemory fields, we can tell them apart more
unambiguously.

Test: Re-ran android/scripts/generate-vulkan-sources.sh, no crashes,
no code changes

Change-Id: Ib20bd6261b9811f34946ca7dc368fe091dfb7afd