Emulate sampler alpha channel for non-alpha textures
Sampler clamp-to-border will fail if the border color set to transparent
while the sampler is used on a software decompressed non-alpha texture,
because internally we decompress textures into RGBA formats. In such
situation we replace the original sampler with one that has opaque
border color.

Test: CTS dEQP-VK.pipeline.sampler.view_type.*

(Some EAC formats are still broken and require follow-up CLs.)

Change-Id: I8c9e6383d5746abec716e4c62d8181a6ad7000ad