[vulkan] Add general transform codegen + external memory transforms
bug: 111137294
bug: 122080810

This CL introduces a general way to transform Vulkan structs on host in
a way that isolates the manually edited parts to the interesting parts,
while removing the boilerplate of creating a new wrapper for an API call
every time something needs to be custom.

This CL includes, as its first example, transformation of external
memory-related structs.

Change-Id: I65a785780935148ff4753c2cec0152491ca35361