[vulkan] Record create info for images/buffers
bug: 122080810

This CL records the basic create info
for each Vulkan image and buffer created on the guest.

This data will be needed if we perform "export"
allocation of external memory with a dedicated VkImage/Buffer.

In this case, we need to retrieve the
image dimensions/format or buffer size,
and calculate AHardwareBuffer_Desc parameters from that.

Change-Id: I0d24e34dbfdf043ec126a8ae4132358fbe29e771