[vulkan] If Vulkan enabled, make ColorBuffers from blob
bug: 122080810

It can be possible to create AHardwareBuffers
with a Vulkan external buffer.
Currently, we rely on there being one ColorBuffer
for each such AHardwareBuffer, except
when blob format is involved.

This CL causes a ColorBuffer to be created
even if the format is blob format.

Then, we get a ColorBuffer ID with which to correlate
Vulkan external buffers.

Change-Id: I8109b15d1e750b72d262c9d7aa33c4726244796f