[vulkan] Block host visible memory from being external memory
bug: 122080810

Since we currently virtualize all host visible memory,
a different, more complicated mechanism would be required for
host visible external memory support in the guest.

Most likely, that would involve having separate allocations for them
on the host as well, or allocated from a separate external memory arena
that on the host (as in guest/host) is host (as in host/device) visible and external.

Change-Id: Iba6de59e64f6e794ae78d34489d6a659eb51abdd