[vulkan] Use C++ FIDL bindings and add Android.bp
Update the code to use C++ FIDL bindings everywhere instead
of using C bindings in some places. No change in behavior.

This also adds an Android.bp file that can be used to build
a libvulkan_goldfish.so for Fuchsia outside the Fuchsia tree.
Tracing is temporarily disabled for this build.

Change-Id: I96b29497cd7b6f4fd6c614d9c653dc5b1830feaf