[vulkan] Fix support for optional strings
Null passed to optional string arguments is currently being
converted to an empty string during serialization. This adds
the appropriate checks to allow null to be passed through to
the host.

The pLayerName argument passed to
vkEnumerateInstanceExtensionProperties is an example where
passing null through to the host is important. It happened
to work today without this fix because the vulkan loader is
converting empty strings to null before calling the ICDs.

Change-Id: I023ae9238764e828ab521f7799b565f848924046