[vulkan] Improve pipe performance on Fuchsia.
This makes use of the new Call method to write and
readback using a single IPC call. The IOStream
interface has been updated to allow for this optimization.

It also implements buffered reads. Each time we readback
data on Fuchsia we attempt to read a larger fixed size,
which we consume until another read is needed.

Change-Id: Ia7aabb5849c651f6d777da996e9d6d271722d0c8