[vulkan] WorkPool (guest)
Generic job scheduler (inspired by libyarn) that lets us
easily have a pool of worker threads specifically for triggering
signals and waits for Vulkan external sync primitives.

The plan is to let this be a quick / dirty version to unblock the bug
and serve as a reference for intended behavior in that use case, then
migrate to libyarn once it gets into aosp.

bug: 139194471
Change-Id: Ida19518250e0ffc05a5eb2924e04b53a65f129e4