Don't use hardware backed gralloc buffer for blob if not asked for
Camera might ask for a CPU buffer and we don't want to back it by GPU
because it might be too big for GL texture. We use a flag
AHARDWAREBUFFER_USAGE_GPU_DATA_BUFFER to distinguish those by requested
by camera versus by Vulkan (which does need to be backed by GPU).

BUG: 141183213
Change-Id: I1be9d586e4f4a0bd362d0212a81df721f826fcc8