goldfish-codecs: Add Goldfish codecs plugin
This will create a vendor/lib{64}/libstagefrighthw.so
and the media framework will load them if exist.

Currently, it has no real codec yet, just empty library.
and it should not impact the current software codecs.

During loading time, ro.kernel.qemu.hwcodec.${codecname}
is checked, and it only tries to load the codec if the
property is set to 1.

Right now those properties are set to 0 by default, and
they will only be set to 1 when the host has the proper
harware codec support.

Note: this CL is a combination of the following cls from
bohu@ and joshuaduong@ respectively:

ag/9406386 and ag/9406405 and ag/9406408

BUG: 124388359
Test: mmm device/generic/goldfish-opengl/system/codecs

Change-Id: I0ca9e9d2d6a044767c097bf2d16971275faf44bd